+++ src/main/Recorder.ts
@@ -8,6 +8,7 @@ import {
   EOBSOutputSignal,
   ERecordingFormat,
   ERecordingState,
+  ESourceFlags,
 } from './obsEnums';
 import { deleteVideo, fixPathWhenPackaged, getSortedVideos } from './util';
 import { IOBSDevice, Metadata, RecStatus, TAudioSourceType } from './types';
@@ -493,6 +494,11 @@ export default class Recorder {
     this.audioInputDevices.forEach((device) => {
       const index = this.audioInputDevices.indexOf(device);
       const channel = this.audioInputChannels[index];
+
+      if (this.cfg.get('obsForceMono')) {
+        device.flags = ESourceFlags.ForceMono;
+      }
+
       this.addAudioSourceOBS(device, channel);
     });