+++ docs/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - [Issue 293](https://github.com/aza547/wow-recorder/issues/293) - Fix a backend bug where reconfiguring would leak audio device references.
 - [Issue 303](https://github.com/aza547/wow-recorder/issues/303) - Fix error handling so that we don't get a blank screen if something goes wrong.
+- [Issue 306](https://github.com/aza547/wow-recorder/issues/306) - Pass up errors returned by OBS signalling.
 
 ## [3.4.0] - 2023-01-22
 ### Added